angular still not recognizing repeat_int, so changed pud model to several inteer fields

+++ puds/models.py
@@ -8,15 +8,15 @@ class Pud(models.Model):
     notification = models.BooleanField(default=False)
 
     priority = models.CharField(max_length=15)
-    priority_int = models.DecimalField(max_digits=1, decimal_places=0, blank=True, default=0)  # low priority
-    duration = models.DecimalField(max_digits=2, decimal_places=0, blank=True, default=0)
+    priority_int = models.IntegerField(blank=True, default=0)  # low priority
+    duration = models.IntegerField(blank=True, default=0)  # takes no time
     is_completed = models.BooleanField(default=False)
 
     repeat = models.CharField(max_length=10, blank=True, default='Perpetual')
-    repeat_int = models.DecimalField(max_digits=1, decimal_places=0, blank=True, default=0)  # perpetual
+    repeat_int = models.IntegerField(blank=True, default=0)  # perpetual
     need_repeat = models.BooleanField(default=False)
 
-    notify_when = models.DecimalField(max_digits=1, decimal_places=0, blank=True, default=0)
+    notify_when = models.IntegerField(blank=True, default=0)
 
     created_at = models.DateTimeField(auto_now_add=True)
     updated_at = models.DateTimeField(auto_now=True)
